#97 - rename logging -> dumping
